Updated button functionality in preparation view. Unified minimap and preparation phase item placement. Made the IconButton control more XAML friendly.
