+++ coh2-battlegrounds/coh2-battlegrounds/Lobby/MVVM/Models/LobbyModel.cs
@@ -29,49 +29,12 @@ using BattlegroundsApp.Modals.Dialogs.MVVM.Models;
 using BattlegroundsApp.MVVM;
 using BattlegroundsApp.MVVM.Models;
 using BattlegroundsApp.Resources;
-using BattlegroundsApp.Utilities;
+
+using static BattlegroundsApp.Lobby.MVVM.Models.LobbyAuxModels;
 
 namespace BattlegroundsApp.Lobby.MVVM.Models;
 
 public abstract class LobbyModel : IViewModel, INotifyPropertyChanged {
-
-    public record LobbyButton(string Title, bool IsEnabled, RelayCommand Click, Visibility Visible, string Tooltip);
-
-    public record LobbyMutButton(RelayCommand Click, Visibility Visible) : INotifyPropertyChanged {
-        private bool m_isEnabled;
-        private Visibility m_iconVisible;
-        private string? m_tooltip;
-        private string? m_title;
-        public event PropertyChangedEventHandler? PropertyChanged;
-        public bool IsEnabled {
-            get => this.m_isEnabled;
-            set {
-                this.m_isEnabled = value;
-                this.PropertyChanged?.Invoke(this, new(nameof(IsEnabled)));
-            }
-        }
-        public string? Tooltip {
-            get => this.m_tooltip;
-            set {
-                this.m_tooltip = value;
-                this.PropertyChanged?.Invoke(this, new(nameof(Tooltip)));
-            }
-        }
-        public string? Title {
-            get => this.m_title;
-            set {
-                this.m_title = value;
-                this.PropertyChanged?.Invoke(this, new(nameof(Title)));
-            }
-        }
-        public Visibility NotificationVisible {
-            get => this.m_iconVisible;
-            set {
-                this.m_iconVisible = value;
-                this.PropertyChanged?.Invoke(this, new(nameof(NotificationVisible)));
-            }
-        }
-    }
 
     public record ScenOp(Scenario Scenario) {
         private static readonly Dictionary<string, string> _cachedNames = new();
