+++ coh2-battlegrounds/coh2-battlegrounds/Lobby/MVVM/Views/LobbyPlanningOverview.xaml.cs
@@ -7,9 +7,9 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Shapes;
 
-using Battlegrounds.Game;
 using Battlegrounds.Game.Database;
 
+using BattlegroundsApp.Controls;
 using BattlegroundsApp.Lobby.MVVM.Models;
 using BattlegroundsApp.Lobby.Planning;
 using BattlegroundsApp.Utilities;
@@ -279,37 +279,7 @@ public partial class LobbyPlanningOverview : UserControl {
         } else if (e.Action is NotifyCollectionChangedAction.Add && e.NewItems is not null) {
 
             foreach (LobbyPlanningMinimapItem item in e.NewItems) {
-
-                // Grab ico
-                string ico = item.EntityBlueprint.Name switch {
-                    "starting_position_shared_territory" => $"Icons_minimap_mm_starting_point_{item.Owner + 1}",
-                    "victory_point" => "Icons_minimap_mm_victory_point",
-                    _ => string.Empty
-                };
-
-                // Bail if no icon is defined
-                if (string.IsNullOrEmpty(ico)) {
-                    continue;
-                }
-
-                // Calculate position
-                var pos = item.Scenario.ToMinimapPosition(this.ScenarioCanvas.Width, this.ScenarioCanvas.Height, item.WorldPos);
-
-                // Create image
-                Image img = new() {
-                    Width = 24,
-                    Height = 24,
-                    Source = App.ResourceHandler.GetIcon("minimap_icons", ico),
-                    RenderTransformOrigin = new(0.5, 0.5)
-                };
-
-                // Add to canvas group
-                this.ScenarioCanvas.Children.Add(img);
-
-                // Display
-                img.SetValue(Canvas.LeftProperty, pos.X - 12);
-                img.SetValue(Canvas.BottomProperty, pos.Y - 12);
-
+                Minimap.AddMinimapItem(this.ScenarioCanvas, this.ScenarioCanvas.Width, this.ScenarioCanvas.Height, item.Scenario, item.WorldPos, item.Owner, item.EntityBlueprint);
             }
 
         }
